backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
redis[hiredis]==5.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
google-genai==1.5.0
langgraph==0.2.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
aiofiles==24.1.0
